python_backend/requirements.txt
# ─────────────────────────────────────────────────────────────
# Core runtime & web
# Pin Flask to the newest 3.x line and let it pull *compatible*
# Werkzeug ≥ 3 and itsdangerous ≥ 2.2, removing the previous
# hard-downgrades that caused the conflict.
flask==3.1.1          # latest stable, Py 3.12 wheels available
httpx>=0.27
pydantic>=2.7
pathspec==0.12.1      # last released version on PyPI
tiktoken>=0.6

# ─────────────────────────────────────────────────────────────
# Smart-Select – embeddings / vector search
sentence-transformers~=3.4
faiss-cpu~=1.8.0
scikit-learn~=1.7

# ─────────────────────────────────────────────────────────────
# Existing runtime / tooling
flask-cors==4.0.0
python-dotenv==1.0.1
# Werkzeug & itsdangerous now come via Flask 3.1.1 (Werkzeug ≥ 3.1,
# itsdangerous ≥ 2.2) – explicit pins removed.
Jinja2==3.1.2
pytest==7.4.0
gunicorn==21.2.0
tree_sitter==0.21.3
tree_sitter_languages==1.10.2 ; python_version < "3.13"
tree_sitter_language_pack>=0.7.0 ; python_version >= "3.13"
tree_sitter_cpp==0.23.4
tree_sitter_c==0.23.4
responses==0.25.0
